+++ Events.lua
@@ -77,25 +77,25 @@ function MyDungeonsBook:COMBAT_LOG_EVENT_UNFILTERED()
 	end
 	if ((subEventPrefix:match("^SPELL") or subEventPrefix:match("^RANGE")) and subEventSuffix == "DAMAGE") then
 		local spellId, _, _, amount, overkill, _, _, _, _, crit = select(12, CombatLogGetCurrentEventInfo());
-		self:TrackAllDamageDoneToPartyMembers(dstName, srcGUID, spellId, amount);
+		self:TrackAllDamageDoneToPartyMembers(srcName, dstName, srcGUID, spellId, amount);
 		self:TrackAllDamageDoneByPartyMembers(srcName, srcGUID, spellId, amount, overkill, crit);
 		self:TrackSLDamageDoneToSpecificUnits(srcName, srcGUID, spellId, amount, overkill, dstName, dstGUID);
 	end
 	if (subEventName == "SWING_DAMAGE") then
 		local amount, overkill, _, _, _, _, crit = select(12, CombatLogGetCurrentEventInfo());
-		self:TrackAllDamageDoneToPartyMembers(dstName, srcGUID, -2, amount);
+		self:TrackAllDamageDoneToPartyMembers(srcName, dstName, srcGUID, -2, amount);
 		self:TrackAllDamageDoneByPartyMembers(srcName, srcGUID, -2, amount, overkill, crit);
 		self:TrackSLDamageDoneToSpecificUnits(srcName, srcGUID, -2, amount, overkill, dstName, dstGUID);
 	end
 	if (subEventName == "SPELL_EXTRA_ATTACKS") then
 		local amount = select(12, CombatLogGetCurrentEventInfo());
-		self:TrackAllDamageDoneToPartyMembers(dstName, srcGUID, -2, amount);
+		self:TrackAllDamageDoneToPartyMembers(srcName, dstName, srcGUID, -2, amount);
 		self:TrackAllDamageDoneByPartyMembers(srcName, srcGUID, -2, amount, 0, false);
 	end
 	if (subEventPrefix:match("^SPELL") and
 		subEventSuffix == "MISSED") then
 		local spellId, _, _, _, _, amount = select(12, CombatLogGetCurrentEventInfo());
-		self:TrackAllDamageDoneToPartyMembers(dstName, srcGUID, spellId, amount);
+		self:TrackAllDamageDoneToPartyMembers(srcName, dstName, srcGUID, spellId, amount);
 	end
 	if (subEventName == "SPELL_AURA_APPLIED" or
 		subEventName == "SPELL_AURA_APPLIED_DOSE") then